Reorganize cascade.dismod a bit
