--- src/cascade/dismod/__init__.py
@@ -1,2 +1,0 @@
-class DismodFileError(Exception):
-    pass
