+++ src/cascade/dismod/db/__init__.py
@@ -1,0 +1,2 @@
+class DismodFileError(Exception):
+    pass
